backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ledgeralps"
version = "0.1.0"
description = "Swiss-compliant accounting and invoicing — local-first"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Validation & serialization
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",

    # Swiss-specific: QR-facture
    "qrbill>=1.0.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",

    # ISO 20022 XML
    "lxml>=5.0.0",
    "xmlschema>=3.0.0",

    # PDF generation
    "weasyprint>=62.0",
    "jinja2>=3.1.4",

    # Utilities
    "python-dateutil>=2.9.0",
    "babel>=2.14.0",
    "pytz>=2024.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
